+++ src/utils/include/utils/sampling.hpp
@@ -50,8 +50,10 @@ get_cylindrical_sampling_positions(std::pair<double, double> const &r_limits,
                                    std::pair<double, double> const &z_limits,
                                    size_t n_r_bins, size_t n_phi_bins,
                                    size_t n_z_bins, double sampling_density) {
-  auto const delta_r = (r_limits.second - r_limits.first) / n_r_bins;
-  auto const delta_phi = (phi_limits.second - phi_limits.first) / n_phi_bins;
+  auto const delta_r =
+      (r_limits.second - r_limits.first) / static_cast<double>(n_r_bins);
+  auto const delta_phi =
+      (phi_limits.second - phi_limits.first) / static_cast<double>(n_phi_bins);
 
   // For the smallest bin we chose samples along the z-direction for a single
   // azimuthal angle per bin such that we fulfill the sampling density
@@ -61,7 +63,8 @@ get_cylindrical_sampling_positions(std::pair<double, double> const &r_limits,
   auto const min_n_samples = std::max(
       n_z_bins,
       static_cast<size_t>(std::round(smallest_bin_volume * sampling_density)));
-  auto const delta_z = (z_limits.second - z_limits.first) / min_n_samples;
+  auto const delta_z =
+      (z_limits.second - z_limits.first) / static_cast<double>(min_n_samples);
 
   auto const r_range =
       make_lin_space(r_limits.first + .5 * delta_r, r_limits.second, n_r_bins,
